finish output
