+++ agarkov.sergey/S1/outputfunc.hpp
@@ -7,9 +7,9 @@
 namespace agarkov
 {
   using pair_t = std::pair< std::string, ForwardList< size_t > >;
-  void outputName(std::ostream& out, ForwardList< pair_t >& lists);
-  void outputNum(std::ostream& out, ForwardList< pair_t >& lists);
-  void outputSum(std::ostream& out, ForwardList< pair_t >& lists);
+  void outputName(std::ostream& out, const ForwardList< pair_t >& lists);
+  void outputNum(std::ostream& out, const ForwardList< pair_t >& lists);
+  void outputSum(std::ostream& out, const ForwardList< pair_t >& lists);
 }
 
 #endif
